backend - v1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
torch==2.0.1
transformers==4.30.2
Pillow==10.0.0
numpy==1.24.3
pdf2image==1.16.3
python-magic==0.4.27
PyMuPDF==1.22.5
tqdm==4.65.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.1.0
httpx==0.24.1